+++ packages/sapa/src/functions/vnode/index.js
@@ -419,7 +419,7 @@ export class VNodeComponent extends VNode {
     );
 
     if (oldId) {
-      // this.instance.setId(oldId);
+      this.instance.setId(oldId);
     }
 
     if (hooks && hooks.__stateHooks?.length) {
@@ -433,7 +433,7 @@ export class VNodeComponent extends VNode {
     }
 
     if (Object.keys(children).length) {
-      // this.instance.setChildren(children);
+      this.instance.setChildren(children);
     }
 
     // 새로운 리소스를 만들었으니 이전 리소스를 제거한다.